Moving adapters to feed component
