--- forex/src/main/java/com/apssouza/mytrade/trading/forex/feed/signal/SignalStream.java
@@ -1,4 +1,0 @@
-package com.apssouza.mytrade.trading.forex.feed.signal;
-
-public interface SignalStream {
-}
